stride: more flexible stride/color conversion
Refactor stride transform element to address a number of limitations:
1) support converting buffers from one rowstride to another, in addition to
just handling conversion from strided <-> unstrided.
2) refactor convert code to make it easier to add new formats
3) refactor caps handling code to build template caps based upon color
formats listed in convert (stride_conversions table).
4) refactor caps parsing/building to correctly handle RGB formats
5) add support for crop.. currently we optimize by just only copying the
uncropped part of the frame, but this is the first step to true handling of
cropping, so that we can crop out padding for the benefit of sink elements
that don't understand crop or stride.  (The convert code handles it fine..
the caps parsing/building in gststridetransform.c would need to handle caps
re-negotiation when the crop changes for this to be complete.)